% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_data.R
\name{att_from_data}
\alias{att_from_data}
\title{Look for functions called in data loading code}
\usage{
att_from_data(chr)
}
\arguments{
\item{chr}{A character vector containing the code as a string. The code should follow the pattern used for loading data with \code{data()}, specifying the dataset and package.}
}
\value{
A character vector containing the names of the packages from which datasets are being loaded.
}
\description{
Look for functions called in data loading code
}
\examples{

vec_char <- 'data("starwars", package = "dplyr")'
att_from_data(vec_char)
}
