% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_description.R, R/att_to_description.R
\name{att_from_description}
\alias{att_from_description}
\alias{att_from_description}
\title{Return all package dependencies from current package}
\usage{
att_from_description(path = "DESCRIPTION", dput = FALSE,
  field = c("Depends", "Imports", "Suggests"))

att_from_description(path = "DESCRIPTION", dput = FALSE,
  field = c("Depends", "Imports", "Suggests"))
}
\arguments{
\item{path}{path to the DESCRIPTION file}

\item{dput}{if FALSE return a vector instead of dput output}

\item{field}{DESCRIPTION fied to parse, Import and Depends by default}

\item{path}{path to the DESCRIPTION file}

\item{dput}{if FALSE return a vector instead of dput output}

\item{field}{DESCRIPTION fied to parse, Import and Depends by default}
}
\value{
A character vector with packages names

A character vector with packages names
}
\description{
Return all package dependencies from current package

Return all package dependencies from current package
}
\examples{
\dontrun{
att_from_description()
}
\dontrun{
att_from_description()
}
}
