% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_functions.R
\name{att_from_functions}
\alias{att_from_functions}
\title{Look for functions called with `::`}
\usage{
att_from_functions(path = "R")
}
\arguments{
\item{path}{path of the directory with R scripts}
}
\value{
a vector
}
\description{
Look for functions called with `::`
}
\details{
Calls from pkg::fun in roxygen skeleton are ignored
}
\examples{
\dontrun{
att_from_functions()
}
}
