% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_namespace.R
\name{att_from_namespace}
\alias{att_from_namespace}
\title{return package dependencies from NAMESPACE file}
\usage{
att_from_namespace(path = "NAMESPACE")
}
\arguments{
\item{path}{path to NAMESPACE file}
}
\value{
a vector
}
\description{
return package dependencies from NAMESPACE file
}
\examples{
\dontrun{
att_from_namespace()
}
}
