% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_rmds.R
\name{att_from_rmd}
\alias{att_from_rmd}
\alias{att_from_qmd}
\title{Get all dependencies from a Rmd file}
\usage{
att_from_rmd(
  path,
  temp_dir = tempdir(),
  warn = -1,
  encoding = getOption("encoding"),
  inside_rmd = NULL,
  inline = TRUE
)

att_from_qmd(
  path,
  temp_dir = tempdir(),
  warn = -1,
  encoding = getOption("encoding"),
  inside_rmd = NULL,
  inline = TRUE
)
}
\arguments{
\item{path}{Path to a Rmd file}

\item{temp_dir}{Path to temporary script from purl vignette}

\item{warn}{-1 for quiet warnings with purl, 0 to see warnings}

\item{encoding}{Encoding of the input file; always assumed to be UTF-8 (i.e.,
this argument is effectively ignored).}

\item{inside_rmd}{Logical or \code{NULL}. Whether the function is being called
from inside a knit session, in which case the actual purl step must be
delegated to an external R process. When \code{NULL} (the default), this is
auto-detected via \code{knitr::opts_knit$get("out.format")}.}

\item{inline}{Logical. Default TRUE. Whether to explore inline code for dependencies.}
}
\value{
vector of character of packages names found in the Rmd
}
\description{
Get all dependencies from a Rmd file
}
\examples{

dummypackage <- system.file("dummypackage",package = "attachment")
# browseURL(dummypackage)
att_from_rmd(path = file.path(dummypackage,"vignettes/demo.Rmd"))

}
