% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_rscripts.R
\name{att_from_rscript}
\alias{att_from_rscript}
\title{Look for functions called with \code{::} and library/requires in one script}
\usage{
att_from_rscript(path)
}
\arguments{
\item{path}{path to R script file}
}
\value{
a vector
}
\description{
Look for functions called with \code{::} and library/requires in one script
}
\details{
Uses the R parser to walk the syntax tree so that occurrences of \code{pkg::fun}
or \code{library()/require()/requireNamespace()/loadNamespace()/use()/getFromNamespace()}
inside string literals or comments are ignored.
Named arguments such as \code{library(package = "pkg")} are supported.
}
\examples{
dummypackage <- system.file("dummypackage",package = "attachment")
# browseURL(dummypackage)

att_from_rscript(path = file.path(dummypackage,"R","my_mean.R"))
}
