% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_rscripts.R
\name{att_from_rscripts}
\alias{att_from_rscripts}
\title{Look for functions called with `::` and library/requires in folder of scripts}
\usage{
att_from_rscripts(path = "R")
}
\arguments{
\item{path}{directory with R scripts inside}
}
\description{
Look for functions called with `::` and library/requires in folder of scripts
}
