% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_vignettes.R
\name{att_from_vignette}
\alias{att_from_vignette}
\title{Get all dependencies from a vignette}
\usage{
att_from_vignette(path)
}
\arguments{
\item{path}{Path to a vignette}
}
\description{
Get all dependencies from a vignette
}
\examples{
\dontrun{
att_from_vignette("vignettes/my-vignette.Rmd")
}
}
