% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_to_description.R
\name{att_to_description}
\alias{att_to_description}
\title{Add packages to description}
\usage{
att_to_description(path = "NAMESPACE", path.d = "DESCRIPTION",
  path.v = "vignettes")
}
\arguments{
\item{path}{path to namespace file}

\item{path.d}{path to description file}

\item{path.v}{path to vignettes}
}
\description{
Add packages to description
}
