% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment-package.R
\docType{package}
\name{attachment-package}
\alias{attachment-package}
\alias{_PACKAGE}
\alias{attachment}
\title{Tools to deal with package dependencies}
\description{
The goal of attachment is to help to deal with package dependencies during package development. It also give useful tools to install or list missing packages used inside Rscripts or Rmds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/attachment}
  \item Report bugs at \url{https://github.com/Thinkr-open/attachment/issues}
}

}
\author{
Vincent Guyader <vincent@thinkr.fr>

Sébastien Rochette <sebastien@thinkr.fr>
}
\keyword{internal}
