% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_remotes.R
\name{complete_remote_branch}
\alias{complete_remote_branch}
\title{Construct a Remote Branch Reference String}
\usage{
complete_remote_branch(remote, remoteref = "HEAD")
}
\arguments{
\item{remote}{A character string representing the name of the remote repository (e.g., \code{"origin"}).}

\item{remoteref}{A character string representing the branch or reference within the remote repository.
Default is \code{"HEAD"}.}
}
\value{
A character string. If \code{remoteref} is \code{"HEAD"}, it returns just the \code{remote} string. If
\code{remoteref} is not \code{"HEAD"}, it returns a string in the format \code{"remote@remoteref"}.
}
\description{
This function constructs a reference string for a remote branch in a version control system like Git.
It returns a formatted string combining the remote name and a specified branch or reference.
}
\keyword{internal}
