% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_renv.R
\name{create_renv_for_dev}
\alias{create_renv_for_dev}
\alias{create_renv_for_prod}
\title{create reproducible environments for your R projects with {renv}}
\usage{
create_renv_for_dev(
  path = ".",
  dev_pkg = c("renv", "devtools", "roxygen2", "usethis", "pkgload", "testthat",
    "remotes", "covr", "attachment", "pak", "dockerfiler"),
  folder_to_include = c("dev/", "data-raw/"),
  output = "renv.lock",
  install_if_missing = TRUE,
  document = TRUE
)

create_renv_for_prod(path = ".", output = "renv.lock.prod")
}
\arguments{
\item{path}{path to your current package source folder}

\item{dev_pkg}{package developpement toolbox you need}

\item{folder_to_include}{folder to scan to detect developpment package}

\item{output}{path and name of the file created, default is `./renv.lock`}

\item{install_if_missing}{Logical. Install missing packages. `TRUE` by default}

\item{document}{Logical. Whether to run `[att_amend_desc()]` before
detecting packages in DESCRIPTION.}
}
\value{
a renv.lock file
}
\description{
Tool to create and maintain renv.lock files.
The idea is to have 2 distinct files, one for development and the other for deployment.
Indeed, although package like {attachment} or {pkgload} must be installed to develop,
they are not necessary in your package/application.
}
\examples{
\dontrun{
create_renv_for_dev()
create_renv_for_dev(dev_pkg = "attachment")
create_renv_for_prod()
}
}
