% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_to_description.R
\name{extract_pkg_info}
\alias{extract_pkg_info}
\title{Internal. Core of find_remotes separated for unit tests}
\usage{
extract_pkg_info(pkgdesc)
}
\arguments{
\item{pkgdesc}{Named list of PackageDescriptions}
}
\description{
Internal. Core of find_remotes separated for unit tests
}
