% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_description.R
\name{install_from_description}
\alias{install_from_description}
\title{Install missing package from DESCRIPTION}
\usage{
install_from_description(
  path = "DESCRIPTION",
  field = c("Depends", "Imports", "Suggests"),
  ...
)
}
\arguments{
\item{path}{path to the DESCRIPTION file}

\item{field}{DESCRIPTION fields to parse, "Depends", "Imports", "Suggests" by default}

\item{...}{Arguments to be passed to \code{\link[utils]{install.packages}}}
}
\value{
Used for side effect. Installs R packages from DESCRIPTION file if missing.
}
\description{
Install missing package from DESCRIPTION
}
\examples{
\dontrun{
dummypackage <- system.file("dummypackage", package = "attachment")
# browseURL(dummypackage)

install_from_description(path = file.path(dummypackage,"DESCRIPTION"))
}
}
