% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_description.R
\name{install_if_missing}
\alias{install_if_missing}
\title{install  packages if missing}
\usage{
install_if_missing(to_be_installed, ...)
}
\arguments{
\item{to_be_installed}{a character vector containing required packages names}

\item{...}{Arguments to be passed to \code{\link[utils]{install.packages}}}
}
\description{
install  packages if missing
}
\examples{
\dontrun{
install_if_missing(c("dplyr","fcuk","rusk"))
}

}
