% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_att_params.R
\name{load_att_params}
\alias{load_att_params}
\title{load_att_params}
\usage{
load_att_params(path_to_yaml = "config_attachment.yaml")
}
\arguments{
\item{path_to_yaml}{character The path to the yaml file}
}
\value{
list A named list of att_amend_desc parameters
}
\description{
load_att_params
}
\examples{
# create a list of parameters and tmp file name
parameter_list <- list(
  pkg_ignore = c("remotes", "i"),
  extra.suggests = c("testthat", "rstudioapi")
)
yaml_path <- paste0(tempfile(pattern = "save_att"), ".yaml")

# save params
save_att_params(param_list = parameter_list,
                path_to_yaml = yaml_path)

# read yaml file
config_params <- load_att_params(path_to_yaml = yaml_path)

# clear created yaml file
unlink(yaml_path)

}
