% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_att_params.R
\name{save_att_params}
\alias{save_att_params}
\title{save_att_params}
\usage{
save_att_params(
  param_list,
  path_to_yaml = "config_attachment.yaml",
  overwrite = FALSE
)
}
\arguments{
\item{param_list}{list A named list of all parameters to save}

\item{path_to_yaml}{character The path to the yaml file}

\item{overwrite}{logical Whether to overwrite the yaml file if it already exists}
}
\value{
character The path to the yaml file
}
\description{
save_att_params
}
\examples{
# create a list of parameters and tmp file name
parameter_list <- list(
  pkg_ignore = c("remotes", "i"),
  extra.suggests = c("testthat", "rstudioapi")
)
yaml_path <- paste0(tempfile(pattern = "save_att"), ".yaml")

# save params
save_att_params(param_list = parameter_list,
                path_to_yaml = yaml_path)

# clear created yaml file
unlink(yaml_path)
}
