% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_from_examples.R
\name{att_from_examples}
\alias{att_from_examples}
\title{Get all packages called in examples from R files}
\usage{
att_from_examples(dir.r = "R")
}
\arguments{
\item{dir.r}{path to directory with R scripts.}
}
\value{
Character vector of packages called with library or require.
}
\description{
Get all packages called in examples from R files
}
\examples{
dummypackage <- system.file("dummypackage",package = "attachment")

# browseURL(dummypackage)
att_from_examples(dir.r = file.path(dummypackage,"R"))
}
