% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_rmds.R
\name{att_from_rmd}
\alias{att_from_rmd}
\title{Get all dependencies from a Rmd file}
\usage{
att_from_rmd(path, temp_dir = tempdir())
}
\arguments{
\item{path}{Path to a Rmd file}

\item{temp_dir}{Path to temporary script from purl vignette}
}
\description{
Get all dependencies from a Rmd file
}
\examples{
\dontrun{
att_from_vignette("vignettes/my-vignette.Rmd")
}
}
