% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_rmds.R
\name{att_from_rmds}
\alias{att_from_rmds}
\title{Get all packages called in vignettes folder}
\usage{
att_from_rmds(path = "vignettes")
}
\arguments{
\item{path}{path to directory with Rmds}
}
\value{
Character vector of packages called with library or require.
{knitr} and {rmarkdown} are added by default to allow building the vignettes
 if the directory contains "vignettes" in the path
}
\description{
Get all packages called in vignettes folder
}
\examples{
\dontrun{
att_from_rmds("vignettes")
}
}
