% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_rscripts.R
\name{att_from_rscript}
\alias{att_from_rscript}
\title{Look for functions called with `::` and library/requires in one script}
\usage{
att_from_rscript(path)
}
\arguments{
\item{path}{path to R script file}
}
\value{
a vector
}
\description{
Look for functions called with `::` and library/requires in one script
}
\details{
Calls from pkg::fun in roxygen skeleton and comments are ignored
}
\examples{
dummypackage <- system.file("dummypackage",package = "attachment")
# browseURL(dummypackage)

att_from_rscript(path = file.path(dummypackage,"R","my_mean.R"))
}
