% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_vignettes.R
\name{att_from_vignettes}
\alias{att_from_vignettes}
\title{Get all packages called in vignettes}
\usage{
att_from_vignettes(path = "vignettes")
}
\arguments{
\item{path}{path to vignettes directory}
}
\description{
Get all packages called in vignettes
}
\examples{
\dontrun{
att_from_vignettes("vignettes")
}
}
