% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_to_description.R
\name{att_to_description}
\alias{att_to_description}
\title{Add packages to description}
\usage{
att_to_description(path = "NAMESPACE", path.d = "DESCRIPTION",
  dir.r = "R", dir.v = "vignettes", document = TRUE)
}
\arguments{
\item{path}{path to namespace file}

\item{path.d}{path to description file}

\item{dir.r}{path to directory with R scripts}

\item{dir.v}{path to vignettes directory}

\item{document}{Run function document of devtools package}
}
\description{
Add packages to description
}
