% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment-deprecated.R
\name{attachment-deprecated}
\alias{attachment-deprecated}
\alias{att_to_description}
\title{Deprecated functions}
\usage{
att_to_description(
  path = ".",
  path.n = "NAMESPACE",
  path.d = "DESCRIPTION",
  dir.r = "R",
  dir.v = "vignettes",
  dir.t = "tests",
  extra.suggests = NULL,
  pkg_ignore = NULL,
  document = TRUE,
  normalize = TRUE,
  inside_rmd = FALSE
)
}
\arguments{
\item{path}{path to the root of the package directory. Default to current directory.}

\item{path.n}{path to namespace file.}

\item{path.d}{path to description file.}

\item{dir.r}{path to directory with R scripts.}

\item{dir.v}{path to vignettes directory. Set to empty (dir.v = "") to ignore.}

\item{dir.t}{path to tests directory. Set to empty (dir.t = "") to ignore.}

\item{extra.suggests}{vector of other packages that should be added in Suggests (pkgdown, covr for instance)}

\item{pkg_ignore}{vector of packages names to ignore.}

\item{document}{Run function roxygenise of roxygen2 package}

\item{normalize}{Logical. Whether to normalize the DESCRIPTION file. See \code{\link[desc:desc_normalize]{desc::desc_normalize()}}}

\item{inside_rmd}{Logical. Whether function is run inside a Rmd,
in case this must be executed in an external R session}
}
\value{
List of functions used for deprecation side effects.
Output depends on the deprecated function.
}
\description{
List of functions deprecated. They will be removed in a future release.
}
\section{Details}{

\tabular{rl}{
\code{att_to_description} \tab is now called \code{att_amend_desc}
so that it is the first function proposed when using auto-completion\cr
}
}

\keyword{internal}
