% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment-package.R
\docType{package}
\name{attachment-package}
\alias{attachment}
\alias{attachment-package}
\title{attachment: Deal with Dependencies}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tools to help manage dependencies during package
    development.  This can retrieve all dependencies that are used in R
    files in the "R" directory, in Rmd files in "vignettes" directory and
    in 'roxygen2' documentation of functions. There is a function to
    update the Description file of your package and a function to create a
    file with the R commands to install all dependencies of your package.
    All functions to retrieve dependencies of R scripts and Rmd files can
    be used independently of a package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/attachment}
  \item Report bugs at \url{https://github.com/Thinkr-open/attachment/issues}
}

}
\author{
\strong{Maintainer}: Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})

Authors:
\itemize{
  \item Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder, funder]
}

}
\keyword{internal}
