% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dependencies_file.R
\name{create_dependencies_file}
\alias{create_dependencies_file}
\title{Create a dependencies.R in the `inst` folder}
\usage{
create_dependencies_file(path = "DESCRIPTION", field = c("Depends",
  "Imports"), to = "inst/dependencies.R")
}
\arguments{
\item{path}{path to the DESCRIPTION file}

\item{field}{DESCRIPTION fied to parse, Import and Depends by default}

\item{to}{path to dependencies.R "inst/dependencies.R" by default}
}
\description{
Create a dependencies.R in the `inst` folder
}
\examples{
\dontrun{
create_dependencies_file()
}
}
