% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_to_description.R
\name{find_remotes}
\alias{find_remotes}
\title{Proposes values for Remotes field for DESCRIPTION file based on your installation}
\usage{
find_remotes(pkg)
}
\arguments{
\item{pkg}{Character. Packages to test for potential non-CRAN installation}
}
\value{
List of non-CRAN packages and code to add in Remotes field in DESCRIPTION
}
\description{
Proposes values for Remotes field for DESCRIPTION file based on your installation
}
\examples{
head(find_remotes(installed.packages()[,1]))
\dontrun{
find_remotes(pkg = c("attachment", "desc", "golem"))
# Find from Description
att_from_description() \%>\% find_remotes()
}
}
